feat: keyExpressions 즐겨찾기 눌렸을 시 login modal 띄우기

+++ components/study/KeyExpression/FavoriteButton.tsx
@@ -1,7 +1,9 @@
+import LoginModal from '@components/common/LoginModal';
 import styled from '@emotion/styled';
 import { FavoriteIcon, favoriteSong } from '@public/assets';
+import { useGetUser } from 'hooks/api';
 import { client } from 'lib/api';
-import React from 'react';
+import React, { useState } from 'react';
 import { mutate } from 'swr';
 
 interface Props {
@@ -12,7 +14,14 @@ interface Props {
 }
 
 function FavoriteButton({ id, isSaved, type, songId }: Props) {
+  const user = useGetUser();
+  const [isLoginModalOpened, setIsLoginModalOpened] = useState(false);
   const handleClick = async (id: number) => {
+    if (!user) {
+      setIsLoginModalOpened(true);
+
+      return;
+    }
     if (isSaved) {
       await client.delete(`/user/vocab/${id}`);
     } else {
@@ -26,7 +35,9 @@ function FavoriteButton({ id, isSaved, type, songId }: Props) {
       type={type}
       src={isSaved ? favoriteSong.src : FavoriteIcon.src}
       onClick={() => handleClick(id)}
-    />
+    >
+      {isLoginModalOpened && <LoginModal setIsLoginModalOpened={setIsLoginModalOpened} />}
+    </Styled.Root>
   );
 }
